Rebased from master with mock server fix; moved test case to client_test execute group
